:bulb: Componentize fitbit_challenges
