+++ fitbit_challenges/frontend/src/BUILD.bazel
@@ -1,0 +1,90 @@
+load("@aspect_bazel_lib//lib:copy_file.bzl", "copy_file")
+load("@aspect_rules_ts//ts:defs.bzl", "ts_project")
+load("@npm//:@tailwindcss/cli/package_json.bzl", tailwindcss_bin = "bin")
+
+exports_files([
+    "logo192.png",
+])
+
+copy_file(
+    name = "globalsdts",
+    src = "//react_library:globalsdts",
+    out = "globals.d.ts",
+    visibility = ["//fitbit_challenges/frontend/src:__subpackages__"],
+)
+
+ts_project(
+    name = "App",
+    srcs = [
+        "App.tsx",
+        ":globalsdts",
+    ],
+    allow_js = True,
+    assets = [
+        "fitbit.png",
+        "logo192.png",
+        "App.src.css",
+        "logo.svg",
+    ],
+    composite = True,
+    declaration = True,
+    out_dir = ".",
+    resolve_json_module = True,
+    transpiler = "tsc",
+    tsconfig = "//fitbit_challenges/frontend:tsconfig",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//:node_modules/@types/react",
+        "//:node_modules/react",
+        "//:node_modules/react-router-dom",
+        "//fitbit_challenges/frontend/src/views:AuthView",
+        "//fitbit_challenges/frontend/src/views:ChallengeView",
+        "//fitbit_challenges/frontend/src/views:ChallengesListingView",
+        "//react_library/NavBar",
+    ],
+)
+
+tailwindcss_bin.tailwindcss(
+    name = "tailwindcss",
+    srcs = [
+        "App.src.css",
+        ":App",
+        "//:node_modules/tailwindcss",
+        "//fitbit_challenges/frontend:tailwindconfig",
+    ],
+    outs = [
+        "App.css",
+    ],
+    args = [
+        "--config fitbit_challenges/frontend/tailwind.config.js",
+        "-i fitbit_challenges/frontend/src/App.src.css",
+        "-o fitbit_challenges/frontend/src/App.css",
+    ],
+    log_level = "debug",
+    visibility = ["//fitbit_challenges/frontend:__subpackages__"],
+)
+
+ts_project(
+    name = "index",
+    srcs = ["index.tsx"],
+    allow_js = True,
+    assets = [
+        "App.css",
+    ],
+    composite = True,
+    declaration = True,
+    out_dir = ".",
+    resolve_json_module = True,
+    transpiler = "tsc",
+    tsconfig = "//fitbit_challenges/frontend:tsconfig",
+    visibility = ["//visibility:public"],
+    deps = [
+        ":App",
+        "//:node_modules/@apollo/client",
+        "//:node_modules/@types/node",
+        "//:node_modules/@types/react",
+        "//:node_modules/@types/react-dom",
+        "//:node_modules/react",
+        "//:node_modules/react-dom",
+    ],
+)
