+++ fitbit_challenges/frontend/src/components/FormButton.test.tsx
@@ -1,3 +1,4 @@
+import '@testing-library/jest-dom'
 import { fireEvent, render, screen } from '@testing-library/react';
 import FormButton from "./FormButton";
