+++ fitbit_challenges/frontend/src/fitbit.png
