--- fitbit_challenges/frontend/src/globals.d.ts
@@ -1,6 +1,0 @@
-declare module "*.png" {
-    const value: any;
-    export = value;
- }
-
- declare module '*.css';
