+++ fitbit_challenges/frontend/src/types/BUILD.bazel
@@ -1,0 +1,14 @@
+load("@aspect_rules_ts//ts:defs.bzl", "ts_project")
+
+ts_project(
+    name = "types",
+    srcs = glob(["*.tsx"]),
+    allow_js = True,
+    composite = True,
+    declaration = True,
+    out_dir = ".",
+    resolve_json_module = True,
+    transpiler = "tsc",
+    tsconfig = "//fitbit_challenges/frontend:tsconfig",
+    visibility = ["//visibility:public"],
+)
