+++ fitbit_challenges/frontend/src/views/AuthView.tsx
@@ -1,6 +1,5 @@
 import { gql } from '@apollo/client/core';
 import { useMutation } from '@apollo/client/react/hooks';
-import PageContainer from '../components/PageContainer';
 
 export const FITBIT_AUTH_MUTATION = gql`
     mutation AuthWithFitbit {
@@ -13,19 +12,13 @@ export const FITBIT_AUTH_MUTATION = gql`
 const AuthView = () => {
     const [startFitbitAuth, { data, loading, error }] = useMutation(FITBIT_AUTH_MUTATION);
 
-    let innerContent = <p></p>;
-    if (loading) innerContent = <p>Loading...</p>;
-    else if (error) innerContent = <p>Error: {error.message}</p>;
+    if (loading) return <p>Loading...</p>;
+    else if (error) return <p>Error: {error.message}</p>;
     else if (data === undefined) {
         startFitbitAuth();
     } else {
         window.location.replace(data.authWithFitbit.url);
     }
-    return (
-        <PageContainer>
-            {innerContent}
-        </PageContainer>
-    )
 }
 
 export default AuthView;
