+++ fitbit_challenges/frontend/src/views/ChallengeView.tsx
@@ -1,46 +1,12 @@
-import { gql } from '@apollo/client/core';
 import { useQuery } from '@apollo/client/react/hooks';
 import { useParams } from 'react-router-dom';
-import PageContainer from '../components/PageContainer';
 import WorkweekHustle from '../components/WorkweekHustle';
 import WeekendWarrior from '../components/WeekendWarrior';
 import BingoChallenge from '../components/BingoChallenge';
 import Activity from '../types/Activity';
 import Challenge, {ChallengeType} from '../types/Challenge';
 import User from '../types/User';
-
-export const FETCH_WORKWEEK_HUSTLE_QUERY = gql`
-    query FetchChallenge($id: Int!) {
-          challenges(id: $id) {
-              id
-              challengeType
-              users {
-                fitbitUserId
-                displayName
-              }
-              createdAt
-              startAt
-              endAt
-              ended
-              sealAt
-              sealed
-              activities {
-                id
-                user
-                createdAt
-                recordDate
-                steps
-                activeMinutes
-                distanceKm
-              }
-          }
-          currentUser {
-            fitbitUserId
-            displayName
-            createdAt
-          }
-      }
-`;
+import { FETCH_WORKWEEK_HUSTLE_QUERY } from '../queries';
 
 type ChallengeViewParams = {
     challengeId: string;
@@ -55,13 +21,12 @@ const ChallengeView = () => {
         {variables: { id }},
     );
 
-    let innerContent = <p></p>;
-    if (loading) innerContent = <p>Loading...</p>;
-    else if (error) innerContent = <p>Error: {error.message}</p>;
+    if (loading) return <p>Loading...</p>;
+    else if (error) return <p>Error: {error.message}</p>;
     else if (data.challenges.length < 1) {
-        innerContent = <p>Error: challenge could not be found!</p>;
+        return <p>Error: challenge could not be found!</p>;
     } else if (data.challenges.length > 1) {
-        innerContent = <p>Error: multiple challenges with that ID were found!</p>
+        return <p>Error: multiple challenges with that ID were found!</p>
     } else {
         const challenges: Challenge[] = data.challenges;
         const challenge = challenges[0];
@@ -69,7 +34,7 @@ const ChallengeView = () => {
         const currentUser: User = data.currentUser;
 
         if (challenge.challengeType === ChallengeType.WeekendWarrior) {
-            innerContent = <WeekendWarrior
+            return <WeekendWarrior
                 id={id}
                 users={challenge.users}
                 startAt={challenge.startAt}
@@ -80,7 +45,7 @@ const ChallengeView = () => {
                 activities={activities}
             />;
         } else if (challenge.challengeType === ChallengeType.WorkweekHustle) {
-            innerContent = <WorkweekHustle
+            return <WorkweekHustle
                 id={id}
                 users={challenge.users}
                 startAt={challenge.startAt}
@@ -91,17 +56,11 @@ const ChallengeView = () => {
                 activities={activities}
             />;
         } else if (challenge.challengeType === ChallengeType.Bingo) {
-            innerContent = <BingoChallenge id={id} currentUser={currentUser} />;
+            return <BingoChallenge id={id} currentUser={currentUser} />;
         } else {
             return <p>Invalid challenge type!</p>
         }
     }
-
-    return (
-        <PageContainer>
-            {innerContent}
-        </PageContainer>
-    )
 }
 
 export default ChallengeView;
