--- fitbit_challenges/frontend/tsconfig.json
@@ -1,19 +1,0 @@
-{
-    "include": ["src"],
-    "exclude": ["node_modules", "**/*.spec.ts"],
-    "compilerOptions": {
-        "lib": ["dom", "dom.iterable", "esnext"],
-        "module": "esnext",
-        "moduleResolution": "bundler",
-        "target": "es2015",
-        "allowJs": true,
-        "allowSyntheticDefaultImports": true,
-        "esModuleInterop": true,
-        "isolatedModules": true,
-        "jsx": "react-jsx",
-        "noFallthroughCasesInSwitch": true,
-        "resolveJsonModule": true,
-        "skipLibCheck": true,
-        "strict": true
-    }
-}
